Squashed commit of the following:
commit 5a9d5e7820e6c80cc8325b12e84f8d0f1a69a7aa
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Mar 2 16:44:13 2021 +0100

    ...

commit 3e05e3c2684aecbb4c0217debbf0f9ed02322c02
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Mar 2 16:00:31 2021 +0100

    ...

commit 3187c5c5dcd91047e8cfe70aa453ab9904d8ff00
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Mar 2 14:57:48 2021 +0100

    ...

commit d78e9fb0730cc7feb9a41562f672c910439685d1
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Mar 2 09:32:28 2021 +0100

    ...

commit ba3be4654e84a742ac8573aba510a6328722c340
Author: Jerzy Jaśkiewicz <jurek.jaskiewicz@gmail.com>
Date:   Mon Mar 1 22:05:31 2021 +0100

    ...

commit 91997cac6c74608167394ba3a9d8784de175d322
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Mon Mar 1 16:56:21 2021 +0100

    ...

commit af969121a3a4012532cc191f41301e701db63553
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Mon Mar 1 12:08:24 2021 +0100

    ...

commit 214c2cecc1e11ddd12a72ab4fd3ac39ea2498e0a
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Fri Feb 26 16:53:13 2021 +0100

    nadal nie działa...

commit 70ba07da70f7c3642a30fdf2fdf77e8ed71574f1
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Fri Feb 26 14:57:20 2021 +0100

    wywałka Windows

commit 0482561834fc78cf5adb9cbc374066fa4f05004e
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Thu Feb 25 09:38:54 2021 +0100

    ...

commit 8023fa658ffe663b201e870cbf3257516b0f39b0
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Thu Feb 25 09:38:29 2021 +0100

    ..

commit 63fef863422ca80da733963c5c63d82bc82402cd
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Wed Feb 24 16:25:30 2021 +0100

    ...

commit 5b0829259a215d7080f32dbbebf3036704df3593
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Wed Feb 24 12:55:23 2021 +0100

    ...

commit ec51c7f65a6a7c41721517466e96bbf8b92f82fe
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Feb 23 11:08:07 2021 +0100

    ...

commit 354409610a20e21f38fadb692e5ad9a420a1d835
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Tue Feb 23 09:58:56 2021 +0100

    ...

commit 3505279e30522a069c3a408adbdf1cc29256b672
Author: Jerzy Jaśkiewicz <jurek.jaskiewicz@gmail.com>
Date:   Mon Feb 22 21:04:46 2021 +0100

    ...

commit 6d92078aea8195dcd1187cc196b08984f4a6a9c0
Author: Jerzy Jaśkiewicz <jurek.jaskiewicz@gmail.com>
Date:   Mon Feb 22 15:46:14 2021 +0100

    ...

commit 6b6c03d2bd5d0a09492cdfa062d638f5f864f4a7
Author: Jerzy Jaśkiewicz <jurek@tvp.pl>
Date:   Wed Feb 17 17:02:59 2021 +0100

    updating TVPlayR

+++ TVPlayR.Player/VideoFormat.h
@@ -36,7 +36,7 @@ namespace TVPlayR {
 		property Rational FrameRate { Rational get() { return _frame_rate; } }
 		property bool IsInterlaced { bool get() { return _native_fomat->interlaced(); } }
 		property FieldModeEnum FieldMode { FieldModeEnum get() { return static_cast<FieldModeEnum>(_native_fomat->field_mode()); } }
-		property int Id { int get() { return _native_fomat->type(); }}
+		property int Id { int get() { return static_cast<int>(_native_fomat->type()); }}
 		String^ FrameNumberToString(int frame_number, bool drop_frame);
 		int StringToFrameNumber(String^ tc);
 		int TimeToFrameNumber(TimeSpan time);
